Implement Vim's `@:` binding.
This repeats the last Ex command. Unlike in Vim, it works only for pure commands,
because `scheduleActionStringForEval` forwards to `pureEval`.
This is arguably a good thing, as it avoids accidentally executing an
irreversible IO command, though it may be also confusing (and the risk should be
rather low, as the last command was already executed shortly before anyway).